requirements.txt
fastapi
uvicorn
numpy
soundfile
torch
transformers
nemo_toolkit[tts]==2.4.0
python-docx
